docker/train/requirements.txt
transformers==4.57.3
unsloth==2025.12.5
datasets>=2.18.0
pyyaml
accelerate
psutil